refactor(onboard): split lib.rs into submodules (<1000 lines)
lib.rs was 2140 lines. Extract the wizard's helper functions into cohesive
submodules, leaving the wizard flow itself in lib.rs:

- prompts.rs: terminal UI primitives (prompt_line, prompt_secret, whoami,
  print_qr_code, fuzzy_select, arrow_select).
- messaging.rs: setup_messaging.
- skills.rs: recommended-skill install flow (+ RecommendedSkill/RECOMMENDED_SKILLS).
- security.rs: device-flow auth, TOTP enrollment/verify, agent SSH key.

lib.rs (~840 lines) keeps OnboardArgs and run_onboard_wizard, which drives the
above. No behavior change.
